siprd/requirements.txt
asgiref==3.4.1
astroid==2.4.2
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
colorama==0.4.4
configparser==5.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0
crayons==0.3.0
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.2
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-braces==1.14.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-heroku==0.3.1
django-oauth-toolkit==1.5.0
django-on-heroku==1.1.2
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
dnspython==2.1.0
docopt==0.6.2
environ==1.0
filelock==3.0.12
graphviz==0.16
gunicorn==20.1.0
idna==3.2
importlib-metadata==1.7.0
isort==5.6.4
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==1.0
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.1.0
platformdirs==2.3.0
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.9.1
pycparser==2.20
pydotplus==2.0.2
PyJWT==2.1.0
pylint==2.6.0
pymongo==3.12.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.59
requests==2.26.0
requests-oauthlib==1.3.0
rest-social-auth==6.0.1
selenium==3.141.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse
toml==0.10.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
virtualenvwrapper-win==1.2.6
webdriver-manager==2.3.0
whitenoise==5.3.0
wrapt==1.12.1
yarg==0.1.9
zipp==3.5.0
